make use of embed for templates
https://github.com/dorkamotorka/goby/issues/1
