+++ internal/generator/generator.go
@@ -1,101 +1,72 @@
 package generator
 
 import (
+	"embed"
+	"io"
 	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
 )
 
-func GenerateGoMain(path string) {
+//go:embed templates/*
+var content embed.FS
+
+func GenerateGoMain(path string) error {
 	outputFile := filepath.Join(path, "main.go")
-
-	// Define the file content
-	content := `package main
-
-//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target bpf program program.bpf.c
-
-import (
-	"os"
-	"os/signal"
-	"syscall"
-	"log"
-
-	"github.com/cilium/ebpf/rlimit"
-)
-
-func main() {
-	// Remove resource limits for kernels <5.11.
-	if err := rlimit.RemoveMemlock(); err != nil {
-		log.Fatal("Removing memlock:", err)
-	}
-
-	// Load the compiled eBPF ELF and load it into the kernel.
-	//var objs programObjects
-	//if err := loadProgramObjects(&objs, nil); err != nil {
-	//	log.Fatal("Loading eBPF objects:", err)
-	//}
-	//defer objs.Close()
-
-	// Here you can now attach eBPF Programs
-	// Check Cilium docs for more functionalities: https://pkg.go.dev/github.com/cilium/ebpf
-	
-	log.Println("eBPF program attached. Press Ctrl+C to exit.")
-
-	// Set up signal handling to cleanly exit
-	stop := make(chan os.Signal, 1)
-	signal.Notify(stop, syscall.SIGINT, syscall.SIGTERM)
-
-	// Keep the program running until a signal is received
-	<-stop
-
-	log.Println("Received signal, exiting...")
-}`
 
 	// Create a new file or overwrite an existing one
 	file, err := os.OpenFile(outputFile, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0666)
 	if err != nil {
-		log.Fatal(err)
+		return err
 	}
 	defer file.Close()
 
-	// Write the content to the file
-	_, err = file.WriteString(content)
+	sourceFile, err := content.Open("templates/main.go")
 	if err != nil {
-		log.Fatal(err)
+		return err
+	}
+	defer sourceFile.Close()
+
+	_, err = io.Copy(file, sourceFile)
+	if err != nil {
+		// Write the content to the file
+		return err
 	}
 
 	log.Println("File 'main.go' created successfully!")
+
+	return nil
 }
 
-func GenerateeBPFProgram(path string) {
+func GenerateeBPFProgram(path string) error {
 	outputFile := filepath.Join(path, "program.bpf.c")
-
-	// Define the C code content
-	content := `//go:build ignore
-
-// Here you write an eBPF program
-
-char _license[] SEC("license") = "GPL";`
 
 	// Create a new eBPF file
 	file, err := os.OpenFile(outputFile, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0666)
 	if err != nil {
-		log.Fatal(err)
+		return err
 	}
 	defer file.Close()
 
-	// Write the content to the file
-	_, err = file.WriteString(content)
+	sourceFile, err := content.Open("templates/program.bpf.c")
 	if err != nil {
-		log.Fatal(err)
+		return err
+	}
+	defer sourceFile.Close()
+	_, err = io.Copy(file, sourceFile)
+	// Write the content to the file
+	if err != nil {
+		return err
 	}
 
 	log.Println("File 'program.bpf.c' created successfully!")
+
+	return nil
 }
 
-// DumpBTF runs the bpftool command to dump BTF information into vmlinux.h
 func DumpBTF(path string) error {
+	// DumpBTF runs the bpftool command to dump BTF information into vmlinux.h
 	outputFile := filepath.Join(path, "vmlinux.h")
 
 	// Create vmlinux.h for writing
@@ -106,7 +77,7 @@ func DumpBTF(path string) error {
 	defer file.Close()
 
 	cmd := exec.Command("bpftool", "btf", "dump", "file", "/sys/kernel/btf/vmlinux", "format", "c")
-	cmd.Stdout = file 
+	cmd.Stdout = file
 	cmd.Stderr = os.Stderr
 
 	err = cmd.Run()
@@ -115,7 +86,7 @@ func DumpBTF(path string) error {
 	}
 
 	// Run the command and return any error
-	return err 
+	return err
 }
 
 // DumpMake creates a Makefile for convenience
@@ -129,38 +100,20 @@ func DumpMake(path string) error {
 	}
 	defer file.Close()
 
-	content := `.PHONY: init generate build run exec
-
-# Only build and run
-run: generate build exec
-
-# Initialize go eBPF project
-init:
-	@go mod init main
-	@go mod tidy
-	@go install github.com/cilium/ebpf/cmd/bpf2go@latest
-
-# Generate eBPF skeleton files
-generate:
-	@go generate
-
-# Build the eBPF object file
-build:
-	@go build
-
-# Run the program
-exec:
-	@sudo ./main
-`
+	sourceFile, err := content.Open("templates/Makefile")
+	if err != nil {
+		return err
+	}
+	defer sourceFile.Close()
 
 	// Write the content to the file
-        _, err = file.WriteString(content)
-        if err != nil {
-                log.Fatal(err)
-        }
+	_, err = io.Copy(file, sourceFile)
+	if err != nil {
+		return err
+	}
 
 	log.Println("File 'Makefile' created successfully!")
 
 	// Run the command and return any error
-	return err 
+	return err
 }
