+++ internal/generator/templates/program.bpf.c
@@ -1,0 +1,5 @@
+//go:build ignore
+
+// Here you write an eBPF program
+
+char _license[] SEC("license") = "GPL";
